Add createAnimation file
