+++ src/createAnimatedComponent/createAnimatedComponent.tsx
@@ -43,7 +43,7 @@ import { PropsFilter } from './PropsFilter';
 import {
   startWebLayoutAnimation,
   tryActivateLayoutTransition,
-} from '../reanimated2/layoutReanimation/web/elementManager';
+} from '../reanimated2/layoutReanimation/web/animationsManager';
 import { configureWebLayoutAnimations } from '../reanimated2/layoutReanimation/web/DOMManager';
 
 const IS_WEB = isWeb();
