+++ src/reanimated2/layoutReanimation/web/animationsManager.ts
@@ -1,241 +1,154 @@
 'use strict';
-import { Animations, AnimationsData, WebEasings, TransitionType } from '.';
-import type {
-  AnimationNames,
-  WebEasingsNames,
-  AnimationCallback,
-  CustomConfig,
-} from '.';
-import { convertAnimationObjectToKeyframes } from './animationParser';
-import type {
-  AnimationData,
-  ReanimatedWebTransformProperties,
-  TransitionData,
-} from './animationParser';
-import type { TransformsStyle } from 'react-native';
 
-import { ReduceMotion } from '../../commonTypes';
+import type { ReanimatedHTMLElement } from '../../js-reanimated';
+import { _updatePropsJS } from '../../js-reanimated';
+import type { AnimationConfig, AnimationNames, CustomConfig } from '.';
+import { Animations } from '.';
+import type { AnimatedComponentProps } from '../../../createAnimatedComponent/utils';
+import { LayoutAnimationType } from '../animationBuilder/commonTypes';
+import type { TransitionData } from './animationParser';
+import type { StyleProps } from '../..//commonTypes';
+import { createAnimationWithExistingTransform } from './createAnimation';
+import {
+  getCallbackFromConfig,
+  getDelayFromConfig,
+  getDurationFromConfig,
+  getEasingFromConfig,
+  getReducedMotionFromConfig,
+  handleEnteringAnimation,
+  handleExitingAnimation,
+  handleLayoutTransition,
+} from './elementManager';
+import { areDOMRectsEqual } from './DOMManager';
 
-import { LinearTransition } from './transition/Linear.web';
-import { SequencedTransition } from './transition/Sequenced.web';
-import { FadingTransition } from './transition/Fading.web';
-import { useReducedMotion } from '../../../reanimated2/hook/useReducedMotion';
-import { insertWebAnimation } from './DOMManager';
-
-// Translate values are passed as numbers. However, if `translate` property receives number, it will not automatically
-// convert it to `px`. Therefore if we want to keep exisitng transform we have to add 'px' suffix to each of translate values
-// that are present inside transform.
-function addPxToTranslate(
-  existingTransform: NonNullable<TransformsStyle['transform']>
+export function startWebLayoutAnimation<
+  ComponentProps extends Record<string, unknown>
+>(
+  props: Readonly<AnimatedComponentProps<ComponentProps>>,
+  element: HTMLElement,
+  animationType: LayoutAnimationType,
+  transitionData?: TransitionData
 ) {
-  type RNTransformProp = (typeof existingTransform)[number];
+  const config =
+    animationType === LayoutAnimationType.ENTERING
+      ? props.entering
+      : animationType === LayoutAnimationType.EXITING
+      ? props.exiting
+      : animationType === LayoutAnimationType.LAYOUT
+      ? props.layout
+      : null;
 
-  if (typeof existingTransform === 'string') {
-    throw new Error('[Reanimated] String transform is unsupported.');
+  if (!config) {
+    return;
   }
 
-  const newTransform = existingTransform.map(
-    (transformProp: RNTransformProp) => {
-      const newTransformProp: ReanimatedWebTransformProperties = {};
-      for (const [key, value] of Object.entries(transformProp)) {
-        if (key.includes('translate')) {
-          // @ts-ignore After many trials we decided to ignore this error - it says that we cannot use 'key' to index this object.
-          // Sadly it doesn't go away after using cast `key as keyof TransformProperties`.
-          newTransformProp[key] = `${value}px`;
-        } else {
-          // @ts-ignore same as above.
-          newTransformProp[key] = value;
-        }
-      }
-      return newTransformProp;
-    }
-  );
+  const isLayoutTransition = animationType === LayoutAnimationType.LAYOUT;
 
-  return newTransform;
-}
+  const initialAnimationName =
+    typeof config === 'function' ? config.name : config.constructor.name;
 
-// In order to keep exisitng transform throughout animation, we have to add it to each of keyframe step.
-function addExistingTransform(
-  newAnimationData: AnimationData,
-  newTransform: ReanimatedWebTransformProperties[]
-) {
-  for (const keyframeStepProperties of Object.values(newAnimationData.style)) {
-    if (!keyframeStepProperties.transform) {
-      // If transform doesn't exist, we add only transform that already exists
-      keyframeStepProperties.transform = newTransform;
-    } else {
-      // We insert existing transformations before ours.
-      Array.prototype.unshift.apply(
-        keyframeStepProperties.transform,
-        newTransform
+  // This prevents crashes if we try to set animations that are not defined.
+  // We don't care about layout transitions since they're created dynamically
+  if (!(initialAnimationName in Animations) && !isLayoutTransition) {
+    if (props.entering) {
+      _updatePropsJS(
+        { visibility: 'initial' },
+        { _component: element as ReanimatedHTMLElement }
       );
     }
-  }
-}
 
-/**
- *  Modifies default animation by preserving transformations that given element already contains.
- *
- * @param animationName Name of the animation to be modified (e.g. `FadeIn`).
- * @param existingTransform Transform values that element already contains.
- * @returns Animation parsed to keyframe string.
- */
-export function createAnimationWithExistingTransform(
-  animationName: string,
-  existingTransform: NonNullable<TransformsStyle['transform']>,
-  layoutTransition?: AnimationData
-) {
-  let newAnimationData;
-
-  if (layoutTransition) {
-    newAnimationData = layoutTransition;
-  } else {
-    if (!(animationName in Animations)) {
-      return '';
-    }
-    newAnimationData = structuredClone(AnimationsData[animationName]);
-  }
-
-  const keyframeName = generateNextCustomKeyframeName();
-
-  newAnimationData.name = keyframeName;
-
-  if (typeof existingTransform === 'string') {
-    throw new Error('[Reanimated] String transform is currently unsupported.');
-  }
-
-  const newTransform = addPxToTranslate(existingTransform);
-
-  addExistingTransform(newAnimationData, newTransform);
-
-  const keyframe = convertAnimationObjectToKeyframes(newAnimationData);
-
-  insertWebAnimation(keyframeName, keyframe);
-
-  return keyframeName;
-}
-
-let customKeyframeCounter = 0;
-
-export function generateNextCustomKeyframeName() {
-  return `REA${customKeyframeCounter++}`;
-}
-
-export function getEasingFromConfig(config: CustomConfig): string {
-  const easingName = (
-    config.easingV !== undefined &&
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    config.easingV!.name in WebEasings
-      ? // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        config.easingV!.name
-      : 'linear'
-  ) as WebEasingsNames;
-
-  return `cubic-bezier(${WebEasings[easingName].toString()})`;
-}
-
-function getRandomDelay(maxDelay = 1000) {
-  return Math.floor(Math.random() * (maxDelay + 1)) / 1000;
-}
-
-export function getDelayFromConfig(config: CustomConfig): number {
-  const shouldRandomizeDelay = config.randomizeDelay;
-
-  const delay = shouldRandomizeDelay ? getRandomDelay() : 0;
-
-  if (!config.delayV) {
-    return delay;
-  }
-
-  return shouldRandomizeDelay
-    ? getRandomDelay(config.delayV)
-    : // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      config.delayV! / 1000;
-}
-
-export function getReducedMotionFromConfig(config: CustomConfig) {
-  if (!config.reduceMotionV) {
-    return useReducedMotion();
-  }
-
-  switch (config.reduceMotionV) {
-    case ReduceMotion.Never:
-      return false;
-    case ReduceMotion.Always:
-      return true;
-    default:
-      return useReducedMotion();
-  }
-}
-
-export function getDurationFromConfig(
-  config: CustomConfig,
-  isLayoutTransition: boolean,
-  animationName: AnimationNames
-): number {
-  const defaultDuration = isLayoutTransition
-    ? 0.3
-    : Animations[animationName].duration;
-
-  return config.durationV !== undefined
-    ? // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      config.durationV! / 1000
-    : defaultDuration;
-}
-
-export function getCallbackFromConfig(config: CustomConfig): AnimationCallback {
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  return config.callbackV !== undefined ? config.callbackV! : null;
-}
-
-/**
- * Creates transition of given type, appends it to stylesheet and returns keyframe name.
- *
- * @param transitionType Type of transition (e.g. LINEAR).
- * @param transitionData Object containing data for transforms (translateX, scaleX,...).
- * @returns Keyframe name that represents transition.
- */
-export function TransitionGenerator(
-  transitionType: TransitionType,
-  transitionData: TransitionData,
-  existingTransform?: NonNullable<TransformsStyle['transform']>
-) {
-  const transitionKeyframeName = generateNextCustomKeyframeName();
-  let transitionObject;
-
-  switch (transitionType) {
-    case TransitionType.LINEAR:
-      transitionObject = LinearTransition(
-        transitionKeyframeName,
-        transitionData
-      );
-      break;
-    case TransitionType.SEQUENCED:
-      transitionObject = SequencedTransition(
-        transitionKeyframeName,
-        transitionData
-      );
-      break;
-    case TransitionType.FADING:
-      transitionObject = FadingTransition(
-        transitionKeyframeName,
-        transitionData
-      );
-      break;
-  }
-
-  if (existingTransform) {
-    return createAnimationWithExistingTransform(
-      '',
-      existingTransform,
-      transitionObject
+    console.warn(
+      "[Reanimated] Couldn't load entering/exiting animation. Current version supports only predefined animations with modifiers: duration, delay, easing, randomizeDelay, wtihCallback, reducedMotion."
     );
+    return;
   }
 
-  const transitionKeyframe =
-    convertAnimationObjectToKeyframes(transitionObject);
+  const transform = (props.style as StyleProps)?.transform;
 
-  insertWebAnimation(transitionKeyframeName, transitionKeyframe);
+  const animationName = transform
+    ? createAnimationWithExistingTransform(initialAnimationName, transform)
+    : initialAnimationName;
 
-  return transitionKeyframeName;
+  const animationConfig: AnimationConfig = {
+    animationName: animationName,
+    duration: getDurationFromConfig(
+      config as CustomConfig,
+      isLayoutTransition,
+      initialAnimationName as AnimationNames
+    ),
+    delay: getDelayFromConfig(config as CustomConfig),
+    easing: getEasingFromConfig(config as CustomConfig),
+    reduceMotion: getReducedMotionFromConfig(config as CustomConfig),
+    callback: getCallbackFromConfig(config as CustomConfig),
+  };
+
+  if (animationConfig.reduceMotion) {
+    if (props.entering) {
+      _updatePropsJS(
+        { visibility: 'initial' },
+        { _component: element as ReanimatedHTMLElement }
+      );
+    }
+
+    return;
+  }
+
+  switch (animationType) {
+    case LayoutAnimationType.ENTERING:
+      handleEnteringAnimation(element, animationConfig);
+      break;
+    case LayoutAnimationType.LAYOUT:
+      // `transitionData` is cast as defined because it is a result of calculations made inside componentDidUpdate method.
+      // We can reach this piece of code only from componentDidUpdate, therefore this parameter will be defined.
+
+      // @ts-ignore This property exists in SequencedTransition
+      (transitionData as TransitionData).reversed = config.reversed
+        ? // @ts-ignore This property exists in SequencedTransition
+          config.reversed
+        : false;
+
+      handleLayoutTransition(
+        element,
+        animationConfig,
+        transitionData as TransitionData,
+        transform
+      );
+      break;
+    case LayoutAnimationType.EXITING:
+      handleExitingAnimation(element, animationConfig);
+      break;
+  }
+}
+
+export function tryActivateLayoutTransition<
+  ComponentProps extends Record<string, unknown>
+>(
+  props: Readonly<AnimatedComponentProps<ComponentProps>>,
+  element: HTMLElement,
+  snapshot: DOMRect
+) {
+  if (!props.layout) {
+    return;
+  }
+
+  const rect = element.getBoundingClientRect();
+
+  if (areDOMRectsEqual(rect, snapshot)) {
+    return;
+  }
+
+  const transitionData: TransitionData = {
+    translateX: snapshot.x - rect.x,
+    translateY: snapshot.y - rect.y,
+    scaleX: snapshot.width / rect.width,
+    scaleY: snapshot.height / rect.height,
+    reversed: false, // This field is used only in `SequencedTransition`, so by default it will be false
+  };
+
+  startWebLayoutAnimation(
+    props,
+    element,
+    LayoutAnimationType.LAYOUT,
+    transitionData
+  );
 }
